+++ cucumber-core/src/main/java/io/cucumber/core/backend/TestCaseState.java
@@ -4,6 +4,7 @@ import org.apiguardian.api.API;
 
 import java.net.URI;
 import java.util.Collection;
+import java.util.Optional;
 
 @API(status = API.Status.STABLE)
 public interface TestCaseState {
@@ -98,4 +99,7 @@ public interface TestCaseState {
      */
     Integer getLine();
 
+    default Optional<Step> geCurrentPickleStep() {
+        return Optional.empty();
+    }
 }